quota: move test helper functions to extra file
